+++ CHANGELOG.md
@@ -10,6 +10,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Simplified Grammar, reduced RAM footprint
 - Dot notation can now be used to access properties of literals, such as
 `"someString".length` or `{foo: 'bar'}.foo`.
+- Transform syntax has changed. Arguments are now passed in parentheses, and
+multiple arguments can be defined. Arguments are no longer limited to object
+literals.
 
 ## [v0.2.0] - 2015-03-02
 ### Added